requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
requests==2.31.0
python-docx==1.1.0
reportlab==4.0.9
PyPDF2==3.0.0
Werkzeug==3.0.1
pymupdf==1.23.8
fpdf==1.7.2
